requirements.txt
asgiref==3.10.0
Django==5.2.7
django-mathfilters==1.0.0
django-simple-history==3.10.1
gunicorn==23.0.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
whitenoise==6.11.0
xlsxwriter==3.2.9
